EventPlanner/transactions/serializers.py
from rest_framework import serializers
from .models import Transaction

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = ['transaction_id', 'price', 'user_id', 'buyer_name', 'buyer_phone', 'date']
